Changed the recipient source code to be a lot more object oriented and flexible. Also made another change to the extensionpage class.
This means it is now possible to actually retrieve parsed name and email data!
The code has become a little less future-proof, because I had to copy some parts of the datasources' code, so if the datasources ever change these bits will have to be changed, too. However I feel the benefits of doing so are far more important.